+++ .gitignore
@@ -7,6 +7,9 @@ __pycache__/
 
 # C extensions
 *.so
+*.pyc
+
+__pycache__/
 
 # Distribution / packaging
 .Python
